DOwnload and modiefied bootstrap file for blog administration.  Mde wholesale changes to dashboard.html to produce an initial cut for CMS site web pages.  Still more work to do in dashboard. html file, need to incorporate it into a single index.html file as per the project 3 spec.  Also need to link it with other functionality, such as logout.
The image files were all temporary placeholders.  SOme of the supporting files are not indirectories they should be in (such as a number of css files), as the links to them did not work when put under the project directory structure.

Handlebars needs to be implemnetd or the dashboard.html file.

The other files were all to support the dashboard file, although the jQuery files may not be needed as the code is being linked to off the internet.  Also the site.js functionality may be merged into the index.js file.